# add findCache for memoryIdx