--- a/idx/memory/find_cache.go
+++ b/idx/memory/find_cache.go
@@ -16,26 +16,26 @@ var (
 	findCacheHit = stats.NewCounterRate32("idx.memory.find-cache.hit")
 	// metric idx.memory.find-cache.miss is a counter of findCache misses
 	findCacheMiss = stats.NewCounterRate32("idx.memory.find-cache.miss")
-
-	findCacheSize            = 1000
-	findCacheInvalidateQueue = 100
-	findCacheBackoff         = time.Minute
 )
 
 type FindCache struct {
-	cache map[uint32]*lru.Cache
-	size  int
 	sync.RWMutex
-	newSeries map[uint32]chan struct{}
-	backoff   map[uint32]time.Time
+	cache               map[uint32]*lru.Cache
+	size                int
+	invalidateQueueSize int
+	backoffTime         time.Duration
+	newSeries           map[uint32]chan struct{}
+	backoff             map[uint32]time.Time
 }
 
-func NewFindCache(size int) *FindCache {
+func NewFindCache(size, invalidateQueueSize int, backoffTime time.Duration) *FindCache {
 	fc := &FindCache{
-		cache:     make(map[uint32]*lru.Cache),
-		size:      size,
-		newSeries: make(map[uint32]chan struct{}),
-		backoff:   make(map[uint32]time.Time),
+		cache:               make(map[uint32]*lru.Cache),
+		size:                size,
+		invalidateQueueSize: invalidateQueueSize,
+		backoffTime:         backoffTime,
+		newSeries:           make(map[uint32]chan struct{}),
+		backoff:             make(map[uint32]time.Time),
 	}
 	return fc
 }
@@ -75,7 +75,7 @@ func (c *FindCache) Add(orgId uint32, pattern string, nodes []*Node) {
 		}
 		c.Lock()
 		c.cache[orgId] = cache
-		c.newSeries[orgId] = make(chan struct{}, findCacheInvalidateQueue)
+		c.newSeries[orgId] = make(chan struct{}, c.invalidateQueueSize)
 		c.Unlock()
 	}
 	cache.Add(pattern, nodes)
@@ -118,7 +118,7 @@ func (c *FindCache) InvalidateFor(orgId uint32, path string) {
 	case ch <- struct{}{}:
 	default:
 		c.Lock()
-		c.backoff[orgId] = time.Now().Add(findCacheBackoff)
+		c.backoff[orgId] = time.Now().Add(c.backoffTime)
 		delete(c.cache, orgId)
 		c.Unlock()
 		for i := 0; i < len(ch); i++ {
@@ -127,7 +127,7 @@ func (c *FindCache) InvalidateFor(orgId uint32, path string) {
 			default:
 			}
 		}
-		log.Infof("memory-idx: findCache invalidate-queue full. Disabling cache for %s. num-cached-entries=%d", findCacheBackoff.String(), cache.Len())
+		log.Infof("memory-idx: findCache invalidate-queue full. Disabling cache for %s. num-cached-entries=%d", c.backoffTime.String(), cache.Len())
 		return
 	}