@Dieterbe 2019-03-11
+1 to what Robert said.
any PR with an impact on performance should show relevant benchmark results.