@woodsaj 2019-03-12
Comparison between master and this branch for relevant Benchmarks
```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkConcurrent4Find/partitioned-8            150287        40939         -72.76%
BenchmarkConcurrent4Find/unPartitioned-8          43931         22132         -49.62%
BenchmarkConcurrent8Find/partitioned-8            201209        37329         -81.45%
BenchmarkConcurrent8Find/unPartitioned-8          59867         19462         -67.49%
BenchmarkConcurrentInsertFind/partitioned-8       412081        42666         -89.65%
BenchmarkConcurrentInsertFind/unPartitioned-8     364459        20785         -94.30%

benchmark                                         old allocs     new allocs     delta
BenchmarkConcurrent4Find/partitioned-8            2324           257            -88.94%
BenchmarkConcurrent4Find/unPartitioned-8          637            187            -70.64%
BenchmarkConcurrent8Find/partitioned-8            2324           257            -88.94%
BenchmarkConcurrent8Find/unPartitioned-8          637            187            -70.64%
BenchmarkConcurrentInsertFind/partitioned-8       11788          258            -97.81%
BenchmarkConcurrentInsertFind/unPartitioned-8     10115          187            -98.15%

benchmark                                         old bytes     new bytes     delta
BenchmarkConcurrent4Find/partitioned-8            95997         27794         -71.05%
BenchmarkConcurrent4Find/unPartitioned-8          34411         18879         -45.14%
BenchmarkConcurrent8Find/partitioned-8            95987         27792         -71.05%
BenchmarkConcurrent8Find/unPartitioned-8          34411         18881         -45.13%
BenchmarkConcurrentInsertFind/partitioned-8       524699        27841         -94.69%
BenchmarkConcurrentInsertFind/unPartitioned-8     484934        18897         -96.10%
```