@Dieterbe 2019-03-12
this needs more comments about the design and implementation.
the backoff/InvalidateQueue stuff in particular is hard to follow.