@Dieterbe 2019-03-12
Can we add a  fast-path for the find-cache disabled case? so it immediately returns out of Add/Purge/.. calls when disabled?