@woodsaj 2019-03-12
> Can we add a fast-path for the find-cache disabled case? so it immediately returns out of Add/Purge/.. calls when disabled?

we already do that.  Caches are per orgId, of `c.cache[orgId]` doesnt exist we immediately return.