@Dieterbe 2019-03-13
> > Can we add a fast-path for the find-cache disabled case? so it immediately returns out of Add/Purge/.. calls when disabled?
>
> we already do that. Caches are per orgId, of c.cache[orgId] doesnt exist we immediately return.

this doesn't seem to be true. Add() will simply add the new cache it doesn't exist yet, so any subsequent calls will see a cache and operate on it.