@woodsaj 2019-03-13
> this doesn't seem to be true. Add() will simply add the new cache if it doesn't exist yet, so any subsequent calls will see a cache and operate on it.

https://github.com/grafana/metrictank/blob/master/idx/memory/find_cache.go#L76-L79