@Dieterbe 2019-03-13
Not sure what you're trying to say here. `t` would be `time.Time{}` there (== unix epoch) so that branch wouldn't be taken. what am i missing?