@woodsaj 2019-03-13
> `t` would be `time.Time{}`

If the cache is disabled, `t` will be a time in the future. https://github.com/grafana/metrictank/blob/master/idx/memory/find_cache.go#L138