@Dieterbe 2019-03-19
future proofing the above URL's:
https://github.com/grafana/metrictank/blob/707e6ce7df0c7cd53f73b75a282212f023ae9af8/idx/memory/find_cache.go#L76-L79
https://github.com/grafana/metrictank/blob/707e6ce7df0c7cd53f73b75a282212f023ae9af8/idx/memory/find_cache.go#L138