@Dieterbe 2019-03-19
we cleared it up in a call: the cache cannot be disabled.